refactor: move cart hash and button helpers to utils
